feat(testing): update specjam to 0.0.7

+++ crates/testing/src/pvm.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct Memory {
-    pub address: u32,
+    pub address: u64,
     pub contents: Vec<u8>,
 }
 
@@ -10,7 +10,7 @@ pub struct Memory {
 pub struct TestInput {
     pub name: String,
     #[serde(alias = "initial-regs")]
-    pub initial_regs: Vec<u32>,
+    pub initial_regs: Vec<u64>,
     #[serde(alias = "initial-pc")]
     pub initial_pc: u32,
     #[serde(alias = "initial-memory")]
@@ -27,7 +27,7 @@ pub struct TestOutput {
     #[serde(alias = "expected-status")]
     pub expected_status: String,
     #[serde(alias = "expected-regs")]
-    pub expected_regs: Vec<u32>,
+    pub expected_regs: Vec<u64>,
     #[serde(alias = "expected-memory")]
     pub expected_memory: Vec<Memory>,
     #[serde(alias = "expected-gas")]
